refactor: keep the main code in main.py file
